pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "photonic-synesthesia"
version = "0.1.0"
description = "AI-driven laser show controller for AlphaTheta XDJ-AZ using LangGraph"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Photonic Synesthesia Team" }
]
keywords = ["dmx", "lighting", "ai", "langgraph", "audio-reactive", "laser", "xdj-az"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio :: Analysis",
    "Topic :: Artistic Software",
]

dependencies = [
    # LangGraph Orchestration
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",

    # Audio Analysis
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    "librosa>=0.10.0",
    "sounddevice>=0.4.6",

    # Beat Tracking (BeatNet has complex deps, madmom as fallback)
    "madmom>=0.16.1",

    # MIDI
    "mido>=1.3.0",
    "python-rtmidi>=1.5.0",

    # Computer Vision
    "opencv-python>=4.8.0",
    "mss>=9.0.0",

    # DMX Control (pyftdi for Enttec Open DMX USB)
    "pyftdi>=0.55.0",

    # Configuration & Validation
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",

    # Utilities
    "structlog>=23.0.0",
    "click>=8.1.0",
]

[project.optional-dependencies]
beatnet = [
    "BeatNet>=1.1.0",
    "torch>=2.0.0",
]
ml = [
    "scikit-learn>=1.3.0",
    "essentia>=2.1b6",
]
web = [
    "fastapi>=0.103.0",
    "uvicorn>=0.23.0",
    "websockets>=11.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "pre-commit>=3.4.0",
]
all = [
    "photonic-synesthesia[beatnet,ml,web,dev]",
]

[project.scripts]
photonic = "photonic_synesthesia.ui.cli:main"
photonic-web = "photonic_synesthesia.ui.web_panel:main"

[project.urls]
Homepage = "https://github.com/toimfortes/laser_show_xdj_az"
Repository = "https://github.com/toimfortes/laser_show_xdj_az"

[tool.hatch.build.targets.wheel]
packages = ["src/photonic_synesthesia"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto"
pythonpath = ["src"]
